Remove support for the ava-3 protocol
