--- test/snapshots/protocol-ava-3.js.md
@@ -1,91 +1,0 @@
-# Snapshot report for `test/protocol-ava-3.js`
-
-The actual snapshot is saved in `protocol-ava-3.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## main() config validation: config may not be an empty object
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## main() config validation: rewrite paths must end in a /
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-> Snapshot 2
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## main() config validation: throw when config contains keys other than 'extensions' or 'rewritePaths'
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## main() config validation: throw when config is not a plain object
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-> Snapshot 2
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-> Snapshot 3
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-> Snapshot 4
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## main() config validation: throw when config.extensions contains duplicates
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## main() config validation: throw when config.extensions contains empty strings
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## main() config validation: throw when config.extensions contains non-strings
-
-> Snapshot 1
-
-    Error {
-      message: 'Unexpected Typescript configuration for AVA. See https://github.com/avajs/typescript/blob/v${pkg.version}/README.md for allowed values.',
-    }
-
-## worker(): load rewritten paths files
-
-> Snapshot 1
-
-    'logged in file.js'
